world_bank_assistant/pyproject.toml
[tool.poetry]
name = "world-bank-assistant"
version = "0.1.0"
description = ""
authors = ["estarguars113 <estarguars113@gmail.com>"]
readme = "README.md"
packages = [{include = "world_bank_assistant"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
pinecone-client = "^2.2.4"
transformers = "^4.34.1"
langchain = "^0.0.325"
torch = "^2.1.0"
langchain-experimental = "^0.0.36"
openai = "^0.28.1"
numpy = "^1.26.1"
pandas = "^2.1.2"
tabulate = "^0.9.0"
sentence-transformers = "^2.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
